lidari

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -7,19 +7,19 @@ package bgu.spl.mics.application.objects;
 public class CloudPoint {
 
     // TODO: Define fields and methods.
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public CloudPoint(int x, int y){
+    public CloudPoint(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX(){
+    public double  getX(){
         return x;
     }
 
-    public int getY(){
+    public double getY(){
         return y;
     }
